requirements.txt
# ── Core & UI ─────────────────────────────────────────
streamlit==1.35.0
streamlit-folium==0.20.0
plotly==5.22.0
pyvis==0.3.2

# ── Data & DB ─────────────────────────────────────────
pyodbc==5.1.0
sqlalchemy==2.0.30
pandas==2.2.2
numpy==1.26.4

# ── ML & Prédiction ───────────────────────────────────
scikit-learn==1.5.0
lightgbm==4.3.0
xgboost==2.0.3


# ── Deep Learning / TFT ───────────────────────────────
torch==2.3.0
pytorch-forecasting==1.1.0
pytorch-lightning==2.2.5

# ── Graph ─────────────────────────────────────────────
networkx==3.3
torch-geometric==2.5.3

# ── Causal AI ─────────────────────────────────────────
dowhy==0.12


# ── Optimisation Multi-Objectifs ──────────────────────
pymoo==0.6.1.3
scipy==1.13.1

# ── MLOps ─────────────────────────────────────────────
mlflow==2.13.2

# ── Utilitaires ───────────────────────────────────────
python-dotenv==1.0.1
pydantic==2.7.1
loguru==0.7.2
joblib==1.4.2

# ── Time Series & Stats ───────────────────────────────
statsmodels>=0.14.0

# ── NLP ──────────────────────────────────────────────
wordcloud>=1.9.0

# ── Fix pkg_resources ─────────────────────────────────
setuptools>=68.0.0
shap>=0.45.0
